requirements.txt
PyQt6>=6.6.0
Pillow>=10.0.0
opencv-python>=4.8.0
onnxruntime>=1.16.0
pyinstaller>=6.0.0
